Linear Operators Related to Matrix Equation Solvers

Linear Operators Related to Matrix Equation Solvers

MatrixEquations.lyapop — Function.

lyapop(A :: AbstractMatrix; disc = false, her = false) -> L::LinearOperator

Define, for an n x n matrix A, the continuous Lyapunov operator L:X -> AX+XA' if disc = false or the discrete Lyapunov operator L:X -> AXA'-X if disc = true. If her = false the Lyapunov operator L:X -> Y maps general square matrices X into general square matrices Y, and the associated M = Matrix(L) is a $n^2 \times n^2$ matrix such that vec(Y) = M*vec(X). If her = true the Lyapunov operator L:X -> Y maps symmetric/Hermitian matrices X into symmetric/Hermitian matrices Y, and the associated M = Matrix(L) is a $n(n+1)/2 \times n(n+1)/2$ matrix such that vec(triu(Y)) = M*vec(triu(X)). For the definitions of the Lyapunov operators see:

M. Konstantinov, V. Mehrmann, P. Petkov. On properties of Sylvester and Lyapunov operators. Linear Algebra and its Applications 312:35–71, 2000.

source

lyapop(A :: AbstractMatrix, E :: AbstractMatrix; disc = false, her = false) -> L::LinearOperator

Define, for a pair (A,E) of n x n matrices, the continuous Lyapunov operator L:X -> AXE'+EXA' if disc = false or the discrete Lyapunov operator L:X -> AXA'-EXE' if disc = true. If her = false the Lyapunov operator L:X -> Y maps general square matrices X into general square matrices Y, and the associated M = Matrix(L) is a $n^2 \times n^2$ matrix such that vec(Y) = M*vec(X). If her = true the Lyapunov operator L:X -> Y maps symmetric/Hermitian matrices X into symmetric/Hermitian matrices Y, and the associated M = Matrix(L) is a $n(n+1)/2 \times n(n+1)/2$ matrix such that vec(triu(Y)) = M*vec(triu(X)). For the definitions of the Lyapunov operators see:

M. Konstantinov, V. Mehrmann, P. Petkov. On properties of Sylvester and Lyapunov operators. Linear Algebra and its Applications 312:35–71, 2000.

source

MatrixEquations.invlyapop — Function.

invlyapop(A :: AbstractMatrix; disc = false, her = false) -> LINV::LinearOperator

Define LINV, the inverse of the continuous Lyapunov operator L:X -> AX+XA' for disc = false or the inverse of the discrete Lyapunov operator L:X -> AXA'-X for disc = true, where A is an n x n matrix. If her = false the inverse Lyapunov operator LINV:Y -> X maps general square matrices Y into general square matrices X, and the associated M = Matrix(LINV) is a $n^2 \times n^2$ matrix such that vec(X) = M*vec(Y). If her = true the inverse Lyapunov operator LINV:Y -> X maps symmetric/Hermitian matrices Y into symmetric/Hermitian matrices X, and the associated M = Matrix(LINV) is a $n(n+1)/2 \times n(n+1)/2$ matrix such that vec(triu(X)) = M*vec(triu(Y)). For the definitions of the Lyapunov operators see:

M. Konstantinov, V. Mehrmann, P. Petkov. On properties of Sylvester and Lyapunov operators. Linear Algebra and its Applications 312:35–71, 2000.

source

invlyapop(A :: AbstractMatrix, E :: AbstractMatrix; disc = false, her = false) -> LINV::LinearOperator

Define LINV, the inverse of the continuous Lyapunov operator L:X -> AXE'+EXA' for disc = false or the inverse of the discrete Lyapunov operator L:X -> AXA'-EXE' for disc = true, where (A,E) is a pair of n x n matrices. If her = false the inverse Lyapunov operator LINV:Y -> X maps general square matrices Y into general square matrices X, and the associated M = Matrix(LINV) is a $n^2 \times n^2$ matrix such that vec(X) = M*vec(Y). If her = true the inverse Lyapunov operator LINV:Y -> X maps symmetric/Hermitian matrices Y into symmetric/Hermitian matrices X, and the associated M = Matrix(LINV) is a $n(n+1)/2 \times n(n+1)/2$ matrix such that vec(triu(X)) = M*vec(triu(Y)). For the definitions of the Lyapunov operators see:

M. Konstantinov, V. Mehrmann, P. Petkov. On properties of Sylvester and Lyapunov operators. Linear Algebra and its Applications 312:35–71, 2000.

source

MatrixEquations.invlyapsop — Function.

invlyapsop(A :: AbstractMatrix; disc = false, her = false) -> LINV::LinearOperator

Define LINV, the inverse of the continuous Lyapunov operator L:X -> AX+XA' for disc = false or the inverse of the discrete Lyapunov operator L:X -> AXA'-X for disc = true, where A is an n x n matrix in Schur form. If her = false the inverse Lyapunov operator LINV:Y -> X maps general square matrices Y into general square matrices X, and the associated M = Matrix(LINV) is a $n^2 \times n^2$ matrix such that vec(X) = M*vec(Y). If her = true the inverse Lyapunov operator LINV:Y -> X maps symmetric/Hermitian matrices Y into symmetric/Hermitian matrices X, and the associated M = Matrix(LINV) is a $n(n+1)/2 \times n(n+1)/2$ matrix such that vec(triu(X)) = M*vec(triu(Y)). For the definitions of the Lyapunov operators see:

M. Konstantinov, V. Mehrmann, P. Petkov. On properties of Sylvester and Lyapunov operators. Linear Algebra and its Applications 312:35–71, 2000.

source

invlyapsop(A :: AbstractMatrix, E :: AbstractMatrix; disc = false, her = false) -> LINV::LinearOperator

Define LINV, the inverse of the continuous Lyapunov operator L:X -> AXE'+EXA' for disc = false or the inverse of the discrete Lyapunov operator L:X -> AXA'-EXE' for disc = true, where (A,E) is a pair of n x n matrices in generalized Schur form. If her = false the inverse Lyapunov operator LINV:Y -> X maps general square matrices Y into general square matrices X, and the associated M = Matrix(LINV) is a $n^2 \times n^2$ matrix such that vec(X) = M*vec(Y). If her = true the inverse Lyapunov operator LINV:Y -> X maps symmetric/Hermitian matrices Y into symmetric/Hermitian matrices X, and the associated M = Matrix(LINV) is a $n(n+1)/2 \times n(n+1)/2$ matrix such that vec(triu(X)) = M*vec(triu(Y)). For the definitions of the Lyapunov operators see:

M. Konstantinov, V. Mehrmann, P. Petkov. On properties of Sylvester and Lyapunov operators. Linear Algebra and its Applications 312:35–71, 2000.

source

MatrixEquations.sylvop — Function.

sylvop(A :: AbstractMatrix, B :: AbstractMatrix; disc = false) -> M::LinearOperator

Define the continuous Sylvester operator M: X -> AX+XB if disc = false or the discrete Sylvester operator M: X -> AXB+X if disc = true.

source

sylvop(A::AbstractMatrix, B::AbstractMatrix, C::AbstractMatrix, D::AbstractMatrix) -> M::LinearOperator

Define the generalized Sylvester operator M: X -> AXB+CXD.

source

MatrixEquations.invsylvop — Function.

invsylvop(A :: AbstractMatrix, B :: AbstractMatrix; disc = false) -> MINV::LinearOperator

Define MINV, the inverse of the continuous Sylvester operator M: X -> AX+XB if disc = false or of the discrete Sylvester operator M: X -> AXB+X if disc = true.

source

invsylvop(A::AbstractMatrix, B::AbstractMatrix, C::AbstractMatrix, D::AbstractMatrix) -> MINV::LinearOperator

Define MINV, the inverse of the generalized Sylvester operator M: X -> AXB+CXD.

source

MatrixEquations.invsylvsop — Function.

invsylvsop(A :: AbstractMatrix, B :: AbstractMatrix; disc = false) -> MINV::LinearOperator

Define MINV, the inverse of the continuous Sylvester operator M: X -> AX+XB if disc = false or of the discrete Sylvester operator M: X -> AXB+X if disc = true, where A and B are in Schur forms.

source

invsylvsop(A::AbstractMatrix, B::AbstractMatrix, C::AbstractMatrix, D::AbstractMatrix; DBSchur = false) -> MINV::LinearOperator

Define MINV, the inverse of the generalized Sylvester operator M: X -> AXB+CXD, with the pairs (A,C) and (B,D) in generalized Schur forms. If DBSchur = true, the pair (D,B) is in generalized Schur form.

source

MatrixEquations.sylvsysop — Function.

sylvsysop(A :: AbstractMatrix, B :: AbstractMatrix, C :: AbstractMatrix, D :: AbstractMatrix) -> M::LinearOperator

Define the operator M: (X,Y) -> [ AX+YB; CX+YD ].

source

MatrixEquations.invsylvsysop — Function.

invsylvsysop(A::AbstractMatrix, B::AbstractMatrix, C::AbstractMatrix, D::AbstractMatrix) -> MINV::LinearOperator

Define MINV, the inverse of the linear operator M: (X,Y) -> [ AX+YB; CX+YD ].

source

MatrixEquations.invsylvsyssop — Function.

invsylvsyssop(A::AbstractMatrix, B::AbstractMatrix, C::AbstractMatrix, D::AbstractMatrix) -> MINV::LinearOperator

Define MINV, the inverse of the linear operator M: (X,Y) -> [ AX+YB; CX+YD ], with the pairs (A,C) and (B,D) in generalized Schur forms.

source

MatrixEquations.trmat — Function.

trmat(n::Int, m::Int) -> M::LinearOperator

Define the transposition operator M: X -> X' for all n x m matrices.

source

trmat(X::AbstractMatrix) -> M::LinearOperator

Define the transposition operator M: X -> X' of all matrices of the size of X.

source