Sylvester Matrix Equation Solvers using Kronecker-product Expansions

MatrixEquations.sylvckr — Function

X = sylvckr(A,B,C)

Solve the continuous Sylvester matrix equation

            AX + XB = C

using the Kronecker product expansion of equations. A and B are square matrices, and A and -B must not have common eigenvalues. This function is not recommended for large order matrices.

MatrixEquations.sylvdkr — Function

X = sylvdkr(A,B,C)

Solve the discrete Sylvester matrix equation

            AXB + X = C

using the Kronecker product expansion of equations. A and B are square matrices, and A and -B must not have common reciprocal eigenvalues. This function is not recommended for large order matrices.

MatrixEquations.gsylvkr — Function

X = gsylvkr(A,B,C,D,E)

Solve the generalized Sylvester matrix equation

            AXB + CXD = E

using the Kronecker product expansion of equations. A, B, C and D are square matrices. The pencils A-λC and D+λB must be regular and must not have common eigenvalues. This function is not recommended for large order matrices.

MatrixEquations.sylvsyskr — Function

sylvsyskr(A,B,C,D,E,F) -> (X,Y)

Solve the Sylvester system of matrix equations

            AX + YB = C
            DX + YE = F

using the Kronecker product expansion of equations. (A,D), (B,E) are pairs of square matrices of the same size. The pencils A-λD and -B+λE must be regular and must not have common eigenvalues. This function is not recommended for large order matrices.

MatrixEquations.dsylvsyskr — Function

dsylvsyskr(A,B,C,D,E,F) -> (X,Y)

Solve the dual Sylvester system of matrix equations

   AX + DY = C
   XB + YE = F

using the Kronecker product expansion of equations. (A,D), (B,E) are pairs of square matrices of the same size. The pencils A-λD and -B+λE must be regular and must not have common eigenvalues. This function is not recommended for large order matrices.